Refactored with MailPackageConcern
